requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
pymongo==4.6.3
python-dotenv==1.0.1
dnspython==2.4.2